# swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/update/{id}": {
      "put": {
        "description": "rota para alterar o nome e senha do usuario, o usuario dele esta logado e pode alterar apenas as proprias informações",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "schema": "bearer"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "string": "any"
                },
                "senha": {
                  "string": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "informações alteradas com sucesso"
          },
          "400": {
            "description": "erro ao alterar"
          },
          "403": {
            "description": "Acesso não autorizado"
          }
        }
      }
    },
...
    "/livros/create": {
      "post": {
        "description": "Rota de administradores, inseri um novo livro ",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "schema": "bearer"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "string": "any"
                },
                "autor": {
                  "string": "any"
                },
                "descricao": {
                  "string": "any"
                },
                "ano":{
                  "number": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/livros/update/:id": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "number": "any"
                },
                "titulo": {
                  "string": "any"
                },
                "autor": {
                  "string": "any"
                },
                "ano": {
                  "number": "any"
                },
                "descricao": {
                  "string": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}